24/03

+++ Programacion/3_Eva/Proyecto04/src/com/project04/model/entities/Personal.java
@@ -51,7 +51,7 @@ public void setOccupation(String occupation) {
     this.occupation = occupation;
 }
 
-public float getId_room() {
+public int getId_room() {
     return id_room;
 }
